Add more relations
