--- src/graphql/TCs/levelTC.js
@@ -1,6 +1,0 @@
-const composeMongoose = require('graphql-compose-mongoose').composeMongoose;
-const Level = require('../../models/level');
-
-const LevelTC = composeMongoose(Level);
-
-module.exports = LevelTC;
